# Conflicts:
#	src/main/java/com/zuhlke/ta/prototype/DummySentimentAnalyzer.java
#	src/main/java/com/zuhlke/ta/prototype/TweetService.java
#	src/main/java/com/zuhlke/ta/prototype/solutions/common/PersistentTweetService.java
#	src/main/java/com/zuhlke/ta/twitterclient/TweetBuffer.java
#	src/main/java/com/zuhlke/ta/twitterclient/TwitterClientRunner.java
#	src/test/java/com/zuhlke/ta/prototype/ImporterTest.java
#	src/test/java/com/zuhlke/ta/prototype/PersistentTweetServiceTest.java
#	src/test/java/com/zuhlke/ta/prototype/inmemory/InMemoryTweetServiceTest.java
#	src/test/java/com/zuhlke/ta/prototype/solutions/inmemory/InMemoryTweetServiceTest.java
#	src/test/java/com/zuhlke/ta/prototype/solutions/mapdb/MapDBTweetServiceTest.java
#	src/test/java/com/zuhlke/ta/twitterclient/TweetBufferTest.java
#	src/test/java/integrationtest/zuhlke/ta/prototype/InMemoryTweetServiceIntegrationTest.java
#	src/test/java/integrationtest/zuhlke/ta/twitterclient/TwitterClientTest.java

+++ src/main/java/com/zuhlke/ta/prototype/SentimentAnalyzer.java
@@ -8,5 +8,5 @@ package com.zuhlke.ta.prototype;
  *
  */
 public interface SentimentAnalyzer {
-    float getSentiment(String text);
+    double getSentiment(String text);
 }
