+++ src/main/java/com/zuhlke/ta/sentiment/TwitterSentimentAnalyzerImpl.java
@@ -70,7 +70,7 @@ public class TwitterSentimentAnalyzerImpl implements SentimentAnalyzer {
 		return ngramFilter.filterNgrams(words);
 	}
 	
-	public float getSentiment(String text) {
+	public double getSentiment(String text) {
 		String[] sentences = sentenceDetector.getSentences(text);
 		float result = 0;
 		for(String sentence: sentences){
